requirements.txt
fastapi
uvicorn
requests
python-dotenv
twilio
sarvamai
google-generativeai
python-multipart